Small tweaks to the codebase

+++ lib/tasks/db.rake
@@ -1,8 +1,11 @@
 require './lib/PostiiConstants'
+
 namespace :db do
   desc "add default error messages to the database"
-  task seed_err_msg: :environment do
-    ErrMsg.delete_all
+  task :seed_err_msg, [:drop_table] => :environment do |t, args|
+    if %w[drop_table reset].include? args[:drop_table]
+      ErrMsg.delete_all
+    end
     PostiiConstants::COMMON_ERRORS.each do |err_code, value|
       ErrMsg.create err_code: err_code,
                     message: value[:message],
